Getting rid of the MXFT_INT and MXFT_UINT types - Part 4
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@277 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_compumotor.c
@@ -1798,7 +1798,7 @@ mxd_compumotor_set_parameter( MX_MOTOR *motor )
 }
 
 MX_EXPORT mx_status_type
-mxd_compumotor_simultaneous_start( int num_motor_records,
+mxd_compumotor_simultaneous_start( long num_motor_records,
 				MX_RECORD **motor_record_array,
 				double *position_array,
 				int flags )
