+++ libMx/d_kohzu_sc.c
@@ -740,7 +740,7 @@ mxd_kohzu_sc_set_parameter( MX_MOTOR *motor )
 }
 
 MX_EXPORT mx_status_type
-mxd_kohzu_sc_simultaneous_start( int num_motor_records,
+mxd_kohzu_sc_simultaneous_start( long num_motor_records,
 				MX_RECORD **motor_record_array,
 				double *position_array,
 				int flags )
