+++ libMx/d_kohzu_sc.h
@@ -46,7 +46,7 @@ MX_API mx_status_type mxd_kohzu_sc_find_home_position( MX_MOTOR *motor );
 MX_API mx_status_type mxd_kohzu_sc_constant_velocity_move( MX_MOTOR *motor );
 MX_API mx_status_type mxd_kohzu_sc_get_parameter( MX_MOTOR *motor );
 MX_API mx_status_type mxd_kohzu_sc_set_parameter( MX_MOTOR *motor );
-MX_API mx_status_type mxd_kohzu_sc_simultaneous_start( int num_motor_records,
+MX_API mx_status_type mxd_kohzu_sc_simultaneous_start( long num_motor_records,
 						MX_RECORD **motor_record_array,
 						double *position_array,
 						int flags );
